+++ src/test/java/user/LoginUserTest.java
@@ -1,12 +1,11 @@
 package user;
 
+import base.BaseTest;
 import client.UserClient;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import io.restassured.RestAssured;
 import io.restassured.response.Response;
 import models.User;
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -14,23 +13,12 @@ import static generators.UserGenerator.*;
 import static org.apache.http.HttpStatus.*;
 import static org.hamcrest.Matchers.*;
 
-public class LoginUserTest {
-    private User user;
-    private UserClient userClient;
+public class LoginUserTest extends BaseTest {
 
     @Before
-    public void setUp() {
-        RestAssured.baseURI = "https://stellarburgers.nomoreparties.site";
+    public void createUser() {
         userClient = new UserClient();
         user = randomUser();
-    }
-
-    @After
-    public void tearDown() {
-        String accessToken = userClient.getUserAccessTokenStep(user);
-        if (accessToken != null) {
-            userClient.deleteUserStep(accessToken);
-        }
     }
 
     @Test
